% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_a_ij.R
\name{calc_a_ij}
\alias{calc_a_ij}
\title{a_ij intermediate function to aid in calculating A matrix}
\usage{
calc_a_ij(i, j, mu_p, mu_aux, p)
}
\arguments{
\item{i}{:= first subscript}

\item{j}{:= second subscript}

\item{mu_p}{:= primary server service rate}

\item{mu_aux}{:= auxiliary server service rate}

\item{p}{:= percentage of customers that are of type primary}
}
\value{
a_ij constant
}
\description{
a_ij intermediate function to aid in calculating A matrix
}
\examples{
i <- 1
j <- 2
mu_p <-.4
mu_aux <-.5
p <-.9
calc_a_ij(i, j, mu_p, mu_aux, p)
}
