% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Am.R
\name{Calc_Am}
\alias{Calc_Am}
\title{Calculating A matrix}
\usage{
Calc_Am(K, s, r, lambda, lambda_aux, lambda_p, mu_p, mu_aux, p)
}
\arguments{
\item{K}{:= truncation parameter}

\item{s}{:= number of auxiliary servers}

\item{r}{:= number of primary servers}

\item{lambda}{:= arrival rate of all customers}

\item{lambda_aux}{:= arrival rate of customers of type auxiliary}

\item{lambda_p}{:= arrival rate of customers of type primary}

\item{mu_p}{:= primary server rate}

\item{mu_aux}{:= auxiliary server rate}

\item{p}{:= percentage of customers that are of type primary}
}
\value{
A_m matrix that represents non-boundary behavior
}
\description{
Calculating A matrix
}
\examples{
K <- 13
r <- 10
s <- 5
lambda <- .2
p <-.9
lambda_aux <- (p-1) * lambda
lambda_p <- p * lambda
mu_p <-.4
mu_aux <-.5
Calc_Am(K,s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p)
}
