% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad.R
\name{quad}
\alias{quad}
\title{quadratic function}
\usage{
quad(a, b, c)
}
\arguments{
\item{a}{:= coefficient of X^2}

\item{b}{:= coefficient of X}

\item{c}{:= constant}
}
\value{
Two X values from resulting quadratic equation
}
\description{
quadratic function
}
\examples{
a <- 2
b <- -1
c <- -3
quad(a,b,c)
}
