% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_results.R
\name{Calc_results}
\alias{Calc_results}
\title{Title}
\usage{
Calc_results(lambda, lambda_aux, r, s, mu_p, mu_aux, K, X_i)
}
\arguments{
\item{lambda}{des}

\item{lambda_aux}{des}

\item{r}{des}

\item{s}{des}

\item{mu_p}{des}

\item{mu_aux}{des}

\item{K}{des}

\item{X_i}{des}
}
\description{
Title
}
